Rename timeout to delayed delivery plus some cleanup
