+++ src/NServiceBus.Transport.Msmq.AcceptanceTests/When_handling_delayed_delivery_errors_fail.cs
@@ -119,19 +119,19 @@
 
             public Task<DateTimeOffset?> Next(CancellationToken cancellationToken = default) => impl.Next(cancellationToken);
 
-            public Task Store(TimeoutItem entity, CancellationToken cancellationToken = default) => impl.Store(entity, cancellationToken);
+            public Task Store(DelayedMessage entity, CancellationToken cancellationToken = default) => impl.Store(entity, cancellationToken);
 
-            public Task<bool> Remove(TimeoutItem entity, CancellationToken cancellationToken = default)
+            public Task<bool> Remove(DelayedMessage entity, CancellationToken cancellationToken = default)
             {
                 throw new Exception("Simulated");
             }
 
-            public Task<bool> IncrementFailureCount(TimeoutItem timeout, CancellationToken cancellationToken = default)
+            public Task<bool> IncrementFailureCount(DelayedMessage timeout, CancellationToken cancellationToken = default)
             {
                 throw new Exception("Simulated");
             }
 
-            public Task<TimeoutItem> FetchNextDueTimeout(DateTimeOffset at, CancellationToken cancellationToken = default) => impl.FetchNextDueTimeout(at, cancellationToken);
+            public Task<DelayedMessage> FetchNextDueTimeout(DateTimeOffset at, CancellationToken cancellationToken = default) => impl.FetchNextDueTimeout(at, cancellationToken);
         }
 
         public class MyMessage : IMessage
