+++ src/NServiceBus.Transport.Msmq/DelayedDelivery/Sql/SqlServerDelayedMessageStore.cs
@@ -5,7 +5,7 @@ namespace NServiceBus
     using System.Data.SqlClient;
     using System.Threading;
     using System.Threading.Tasks;
-    using Transport.Msmq.Timeouts;
+    using Transport.Msmq.DelayedDelivery.Sql;
 
     //TODO: Either we should expect that the connection created by the factory is open of we should make it non-async
     /// <summary>
@@ -53,7 +53,7 @@ namespace NServiceBus
         }
 
         /// <inheritdoc />
-        public async Task Store(TimeoutItem timeout, CancellationToken cancellationToken = default)
+        public async Task Store(DelayedMessage timeout, CancellationToken cancellationToken = default)
         {
             using (var cn = await createSqlConnection(cancellationToken).ConfigureAwait(false))
             using (var cmd = new SqlCommand(insertCommand, cn))
@@ -62,7 +62,7 @@ namespace NServiceBus
                 cmd.Parameters.AddWithValue("@destination", timeout.Destination);
                 cmd.Parameters.AddWithValue("@time", timeout.Time);
                 cmd.Parameters.AddWithValue("@headers", timeout.Headers);
-                cmd.Parameters.AddWithValue("@state", timeout.State);
+                cmd.Parameters.AddWithValue("@state", timeout.Body);
                 await cn.OpenAsync(cancellationToken).ConfigureAwait(false);
                 _ = await cmd.ExecuteNonQueryAsync(cancellationToken).ConfigureAwait(false);
             }
@@ -70,7 +70,7 @@ namespace NServiceBus
 
 
         /// <inheritdoc />
-        public async Task<bool> Remove(TimeoutItem timeout, CancellationToken cancellationToken = default)
+        public async Task<bool> Remove(DelayedMessage timeout, CancellationToken cancellationToken = default)
         {
             using (var cn = await createSqlConnection(cancellationToken).ConfigureAwait(false))
             using (var cmd = new SqlCommand(removeCommand, cn))
@@ -83,7 +83,7 @@ namespace NServiceBus
         }
 
         /// <inheritdoc />
-        public async Task<bool> IncrementFailureCount(TimeoutItem timeout, CancellationToken cancellationToken = default)
+        public async Task<bool> IncrementFailureCount(DelayedMessage timeout, CancellationToken cancellationToken = default)
         {
             using (var cn = await createSqlConnection(cancellationToken).ConfigureAwait(false))
             using (var cmd = new SqlCommand(bumpFailureCountCommand, cn))
@@ -128,9 +128,9 @@ namespace NServiceBus
         }
 
         /// <inheritdoc />
-        public async Task<TimeoutItem> FetchNextDueTimeout(DateTimeOffset at, CancellationToken cancellationToken = default)
+        public async Task<DelayedMessage> FetchNextDueTimeout(DateTimeOffset at, CancellationToken cancellationToken = default)
         {
-            TimeoutItem result = null;
+            DelayedMessage result = null;
             using (var cn = await createSqlConnection(cancellationToken).ConfigureAwait(false))
             using (var cmd = new SqlCommand(fetchCommand, cn))
             {
@@ -141,13 +141,13 @@ namespace NServiceBus
                 {
                     if (await reader.ReadAsync(cancellationToken).ConfigureAwait(false))
                     {
-                        result = new TimeoutItem
+                        result = new DelayedMessage
                         {
                             Id = (string)reader[0],
                             Destination = (string)reader[1],
                             Time = (DateTime)reader[2],
                             Headers = (byte[])reader[3],
-                            State = (byte[])reader[4],
+                            Body = (byte[])reader[4],
                             NumberOfRetries = (int)reader[5]
                         };
                     }
